backend/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
astunparse==1.6.3
beautifulsoup4==4.12.3
certifi==2025.8.3
charset-normalizer==3.4.2
distro==1.9.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.7.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
jiter==0.10.0
keras==3.11.3
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
numpy==2.3.2
openai==1.98.0
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
protobuf==6.32.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-decouple==3.8
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.3
rich==14.1.0
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tokenizers==0.21.4
tqdm==4.67.1
transformers==4.55.4
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
